requirements.txt
numpy
torch
torchvision
tensorboardX
mlflow
pillow
matplotlib
dvc
boto3
grafana-client